+++ generators/controller/index.js
@@ -27,6 +27,7 @@
 	const authorEmail    = require('../common/inputs/authorEmail.js');
 	const controllerName = require('./inputs/controllerName.js');
 	const controllerPath = require('./inputs/controllerPath.js');
+	const fileName       = require('./inputs/fileName.js');
 
 	const writingControllerFile = require('./writing/controllerFile.js');
 
@@ -93,6 +94,10 @@
 			return controllerPath.set(this);
 		}
 
+		inputFileName() {
+			return fileName.set(this);
+		}
+
 		writing() {
 			this.now     = moment();
 			this.nowDate = this.now.format('DD/MM/YYYY');
@@ -114,7 +119,7 @@
 
 		end() {
 			logs.priority(this, 'end', false);
-			this.log(chalk.hex(colors.get('green'))('Your controller file is ready ;)'));
+			this.log(chalk.hex(colors.get('green'))('Your controller is ready ;)'));
 		}
 	};